Use vector to hold objects to destroy
ApiObjectList::Destroy() moved the objects it is going to destoy to a
temporary LinkedList. ApiObjectList::Untrack() returns true if the
object is in any list. If another thread runs
ApiObjectBase::DeleteThis() at the same time, that sees the objects are
in a list and proceeds to remove it from list and then destroy it. This
can result in object being removed from list and having DestroyImpl()
called on it twice.

This is fixed by using a std::vector to track the objects to destroy
instead. This avoids the double remove/destruction as
ApiObjectList::Destroy() removes all objects from the original list
under the lock.

Note there is still a problem where the if the last reference is dropped
on another thread before ApiObjectList calls DestroyImpl(). The other
thread will delete the object resulting in DestroyImpl() being a
use-after-free. That is going to fixed as a follow up.

Bug: 429112750, 396294899
Change-Id: I9e8a809c653a1565d1dfbb4a7be48c91a644f540
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/251094
Commit-Queue: Kyle Charbonneau <kylechar@google.com>
Reviewed-by: Loko Kung <lokokung@google.com>
(cherry picked from commit f4cc537192cae225dc058af03d3b7cb45c9d8a7c)
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/253215